+++ packages/react-form-devtools/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @tanstack/react-form-devtools
 
+## 0.2.4
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/form-devtools@0.2.4
+
+## 0.2.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/form-devtools@0.2.3
+
 ## 0.2.2
 
 ### Patch Changes